requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
python-nmap==0.7.1
dnspython==2.6.1
jinja2==3.1.3
email-validator==2.1.0.post1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
psutil==5.9.8 